move group mutations to mutations.ts
