+++ apps/dashboard/src/app/group/modals/create-group-modal.tsx
@@ -1,7 +1,7 @@
 import { type ContextModalProps, modals } from "@mantine/modals"
 import type { FC } from "react"
-import { useCreateGroupMutation } from "../mutations/use-create-group-mutation"
 import { useGroupWriteForm } from "../write-form"
+import { useCreateGroupMutation } from "../mutations"
 
 export const CreateGroupModal: FC<ContextModalProps> = ({ context, id }) => {
   const close = () => context.closeModal(id)
