+++ apps/dashboard/src/app/group/mutations.ts
@@ -1,7 +1,102 @@
-import { useQueryGenericMutationNotification } from "@/lib/notifications"
+import { useQueryGenericMutationNotification, useQueryNotification } from "@/lib/notifications"
 import { useTRPC } from "@/lib/trpc"
 
 import { useMutation, useQueryClient } from "@tanstack/react-query"
+
+export const useCreateGroupMutation = () => {
+  const trpc = useTRPC()
+  const queryClient = useQueryClient()
+  const notification = useQueryNotification()
+
+  return useMutation(
+    trpc.group.create.mutationOptions({
+      onMutate: () => {
+        notification.loading({
+          title: "Oppretter gruppe...",
+          message: "Gruppen blir opprettet.",
+        })
+      },
+      onSuccess: async (data) => {
+        notification.complete({
+          title: "Gruppen er opprettet",
+          message: `Gruppen "${data.name}" har blitt oppdatert.`,
+        })
+
+        await queryClient.invalidateQueries(trpc.group.all.queryOptions())
+      },
+      onError: (err) => {
+        notification.fail({
+          title: "Feil oppsto",
+          message: `En feil oppsto under opprettelsen: ${err.toString()}.`,
+        })
+      },
+    })
+  )
+}
+
+export const useDeleteGroupMutation = () => {
+  const trpc = useTRPC()
+  const queryClient = useQueryClient()
+  const notification = useQueryNotification()
+
+  return useMutation(
+    trpc.group.delete.mutationOptions({
+      onMutate: () => {
+        notification.loading({
+          title: "Sletter gruppen",
+          message: "Gruppen slettes. Vennligst vent.",
+        })
+      },
+      onSuccess: async () => {
+        notification.complete({
+          title: "Gruppen er slettet",
+          message: "Gruppen er fjernet fra systemet.",
+        })
+
+        await queryClient.invalidateQueries(trpc.group.all.queryOptions())
+      },
+      onError: (err) => {
+        notification.fail({
+          title: "Feil oppsto",
+          message: `En feil oppsto under slettingen: ${err.toString()}.`,
+        })
+      },
+    })
+  )
+}
+
+export const useUpdateGroupMutation = () => {
+  const trpc = useTRPC()
+  const queryClient = useQueryClient()
+  const notification = useQueryNotification()
+
+  return useMutation(
+    trpc.group.update.mutationOptions({
+      onMutate: () => {
+        notification.loading({
+          title: "Oppdaterer",
+          message: "Gruppen blir oppdatert.",
+        })
+      },
+      onSuccess: async (data) => {
+        notification.complete({
+          title: "Oppdatert",
+          message: `Gruppen "${data.name}" har blitt oppdatert.`,
+        })
+
+        await queryClient.invalidateQueries(trpc.group.get.queryOptions(data.slug))
+        await queryClient.invalidateQueries(trpc.group.all.queryOptions())
+      },
+      onError: (err) => {
+        notification.fail({
+          title: "Feil oppsto",
+          message: `En feil oppsto under oppdatering: ${err.toString()}.`,
+        })
+      },
+    })
+  )
+}
+
 
 export const useCreateGroupRoleMutation = () => {
   const trpc = useTRPC()
